+++ src/components/Bank.js
@@ -1,11 +1,67 @@
-// function Bank() {
 import React from 'react';
+import TaskList from './Tasks';
+import { AddTaskForm } from './TaskForms';
 
 export default class Bank extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      tasks: props.tasks
+    }
+  }
+
+  componentDidMount() {
+    console.log("Mounted!");
+
+    fetch('./tasks.json').then((res) => res.json()).then((data) => {
+      this.setState((currState) => {
+        return {tasks: data}
+      });
+    });
+  }
+
+  toggleTask = (taskId) => {
+    this.setState((currState) => {
+      let newUpdatedTasksArray = currState.tasks.map((eachTask) => {
+        let taskCopy = Object.assign({}, eachTask)
+        if (taskCopy.id === taskId) {
+          taskCopy.complete = !taskCopy.complete;
+        }
+        return taskCopy;
+      });
+      return {tasks: newUpdatedTasksArray};
+    })
+  }
+
+  addTask = (newDescription) => {
+    this.setState((currState) => {
+      let newTask = {
+        id: currState.tasks.length+1,
+        description: newDescription,
+        complete: false
+      }
+
+      let newTasksArray = currState.tasks.concat( [newTask] )
+
+      return {tasks: newTasksArray};
+    })
+  }
+
+
   render() {
+    console.log("Rendering Bank...");
+
+    let incompleteArray = this.state.tasks.filter((task) => !task.complete);
+    console.log("Number of incomplete tasks", incompleteArray.length);
+
     return(
-      <div>
-          <p>Bank</p>
+      <div className="container">
+        <p className="lead" onClick={() => this.addTask("testing")}>
+          Num things I have to do: <strong>{incompleteArray.length}</strong>
+        </p>
+        <TaskList tasks={this.state.tasks} whatToDoWhenClicked={this.toggleTask} />
+        <AddTaskForm howToAdd={this.addTask} />
       </div>
     )
   }
